+++ package.json
@@ -16,10 +16,10 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "test": "mocha public_html/test/**/*.js",
+    "test": "mocha test/**/*.js",
     "test:unit": "vitest",
     "lint": "eslint . --fix",
-    "format": "prettier --write public_html/src/ public_html/test/"
+    "format": "prettier --write src/ test/"
   },
   "dependencies": {
     "chai": "^4.3.0",
