+++ public/js/app/PasswordNotificator.js
