+++ public/js/app/comparator/ScoreDataFeeHoursAndLockHoursComparator.js
