+++ public/js/app/model/PasswordRegistrationCollection.js
