+++ public/js/app/ui/JQPasswordRegistration.js
