+++ public/js/app/ui/PagePasswordDialog.js
