+++ public/js/app/ui/PagePasswordTrainer.js
