+++ public/js/util/hash/SaltGenerator.js
