--- public_html/gen-manifest-appcache.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-echo "CACHE MANIFEST"
-echo ""
-echo "CACHE:"
-echo '# '$( date +'%Y-%m-%d %H:%M:%S' )
-find -L -type f | grep '\.css$\|\.htm$\|\.html$\|\.ico$\|\.js\|\.map\|\.png$' | sort | grep -oP '(?<=^\./).*$'
