--- public_html/index_build.html
@@ -1,126 +1,0 @@
-<!-- vi: ts=4 et
--->
-<!DOCTYPE html>
-<html manifest="manifest.appcache">
-    <head>
-        <title>PasswordTrainer</title>
-        <meta charset="UTF-8"/>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-
-        <link rel="apple-touch-icon-precomposed" href="img/icon.png"/>
-
-        <link rel="stylesheet" type="text/css" href="libs/jquery-mobile/jquery.mobile.css"/>
-        <link rel="stylesheet" type="text/css" href="css/JQPasswordInput.css"/>
-
-        <!-- js libs -->
-        <script src="libs/crypto-js/core.js"></script>
-        <script src="libs/crypto-js/md5.js"></script>
-        <script src="libs/jquery/jquery.js"></script>
-        <script src="libs/jquery-mobile/jquery.mobile.js"></script>
-
-        <!-- PasswordTrainer js bundle -->
-        <script src="js/bundle.js"></script>
-    </head>
-    <body>
-        <div data-role="page" id="pageTrainPasswords">
-            <div data-role="header" data-position="fixed">
-                <h1>Train passwords</h1>
-            </div>
-
-            <div data-role="content">
-                <div class="ui-field-contain">
-                    <label for="select-password" class="select">Select password:</label>
-                    <select name="select-password" id="select-password"></select>
-                </div>
-
-                <div id="passwordtrainer">
-                    <h2 class="passworddescription"></h2>
-                    <div>
-                        <input type="password" class="password" data-role="none"/><br>
-                        <span class="passwordscore"/>
-                    </div>
-                </div>                
-            </div>
-
-            <div data-role="footer" data-position="fixed">
-                <div data-role="navbar">
-                    <ul>
-                        <li><a href="#pageTrainPasswords" class="ui-btn-active ui-state-persist">Train</a></li>
-                        <li><a href="#pageManagePasswords">Manage</a></li>
-                        <li><a href="#pageImportExport">Import/Export</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-
-        <div data-role="page" id="pageManagePasswords">
-            <div data-role="header" data-position="fixed">
-                <h1>Manage Passwords</h1>
-            </div>
-
-            <div data-role="content">
-                <table class="passwordTable">
-                    <tr>
-                        <th>Description</th>
-                        <th>Score / Level</th>
-                        <th>Status</th>
-                    </tr>
-                </table>
-                <a href="#pagePasswordDialog" data-rel="dialog">Add/Edit password</a>
-            </div>
-
-            <div data-role="footer" data-position="fixed">
-                <div data-role="navbar">
-                    <ul>
-                        <li><a href="#pageTrainPasswords">Train</a></li>
-                        <li><a href="#pageManagePasswords" class="ui-btn-active ui-state-persist">Manage</a></li>
-                        <li><a href="#pageImportExport">Import/Export</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-
-        <div data-role="page" id="pageImportExport">
-            <div data-role="header" data-position="fixed">
-                <h1>Import/Export</h1>
-            </div>
-
-            <div data-role="content">
-                <textarea id="importexportfield"></textarea>
-            </div>
-
-            <div data-role="footer" data-position="fixed">
-                <div data-role="navbar">
-                    <ul>
-                        <li><a href="#pageTrainPasswords">Train</a></li>
-                        <li><a href="#pageManagePasswords">Manage</a></li>
-                        <li><a href="#pageImportExport" class="ui-btn-active ui-state-persist">Import/Export</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-
-        <div data-role="page" id="pagePasswordDialog">
-            <div data-role="header" data-position="fixed">
-                <h1>Add/Edit Passwords</h1>
-            </div>
-
-            <div data-role="content">
-                <table id="passwordregistration">
-                    <tr>
-                        <th>Description:</th>
-                        <td><input type="text" data-role="none" class="passworddescription"/></td>
-                    </tr>
-                    <tr>
-                        <th>Password:</th>
-                        <td><input type="text" data-role="none" class="password"/></td>
-                    </tr>
-                    <tr>
-                        <th>Repeat:</th>
-                        <td><input type="text" data-role="none" class="passwordrepeat"/></td>
-                    </tr>
-                </table>
-            </div>
-        </div>
-    </body>
-</html>
