+++ test/js/app/comparator/TestScoreDataFeeHoursAndLockHoursComparator.js
@@ -1,7 +1,7 @@
 // vi: ts=4 et
 
 const assert = require('chai').assert;
-const ScoreDataFeeHoursAndLockHoursComparator = require('../../../../js/app/comparator/ScoreDataFeeHoursAndLockHoursComparator.js').ScoreDataFeeHoursAndLockHoursComparator
+const ScoreDataFeeHoursAndLockHoursComparator = require('../../../../public/js/app/comparator/ScoreDataFeeHoursAndLockHoursComparator.js').ScoreDataFeeHoursAndLockHoursComparator
 
 describe("ScoreDataFeeHoursAndLockHoursComparator", function() {
     beforeEach( function() {
