+++ test/js/util/hash/TestMD5PasswordHasher.js
@@ -1,8 +1,8 @@
 // vi: ts=4 et
 
 const assert = require('chai').assert;
-const SaltGenerator = require('../../../../js/util/hash/SaltGenerator.js').SaltGenerator;
-const MD5PasswordHasher = require('../../../../js/util/hash/MD5PasswordHasher.js').MD5PasswordHasher;
+const SaltGenerator = require('../../../../public/js/util/hash/SaltGenerator.js').SaltGenerator;
+const MD5PasswordHasher = require('../../../../public/js/util/hash/MD5PasswordHasher.js').MD5PasswordHasher;
 
 describe("MD5PasswordHasher", function() {
     beforeEach(function() {
