+++ test/js/util/json/TestJSONFormatter.js
@@ -1,7 +1,7 @@
 // vi: ts=4 et
 
 const assert = require('chai').assert;
-const JSONFormatter = require('../../../../js/util/json/JSONFormatter.js').JSONFormatter;
+const JSONFormatter = require('../../../../public/js/util/json/JSONFormatter.js').JSONFormatter;
 
 describe("JSONFormatter", function() {
     beforeEach(function() {
