+++ vitest.config.js
@@ -7,7 +7,7 @@ export default mergeConfig(
   defineConfig({
     test: {
       environment: 'jsdom',
-      include: [...configDefaults.include, 'public_html/test/**/*.js'],
+      include: [...configDefaults.include, 'test/**/*.js'],
       exclude: [...configDefaults.exclude, 'e2e/**'],
       root: fileURLToPath(new URL('./', import.meta.url)),
     },
